+++ replica/database.cc
@@ -451,32 +451,6 @@ void backlog_controller::update_controller(float shares) {
 
 namespace replica {
 
-dirty_memory_manager::dirty_memory_manager(replica::database& db, size_t threshold, double soft_limit, scheduling_group deferred_work_sg)
-    : _db(&db)
-    , _region_group("memtable (unspooled)", dirty_memory_manager_logalloc::reclaim_config{
-            .unspooled_hard_limit = threshold / 2,
-            .unspooled_soft_limit = threshold * soft_limit / 2,
-            .real_hard_limit = threshold,
-            .start_reclaiming = std::bind_front(&dirty_memory_manager::start_reclaiming, this)
-      }, deferred_work_sg)
-    , _flush_serializer(1)
-    , _waiting_flush(flush_when_needed()) {}
-
-void
-dirty_memory_manager::setup_collectd(sstring namestr) {
-    namespace sm = seastar::metrics;
-
-    _metrics.add_group("memory", {
-        sm::make_gauge(namestr + "_dirty_bytes", [this] { return real_dirty_memory(); },
-                       sm::description("Holds the current size of a all non-free memory in bytes: used memory + released memory that hasn't been returned to a free memory pool yet. "
-                                       "Total memory size minus this value represents the amount of available memory. "
-                                       "If this value minus unspooled_dirty_bytes is too high then this means that the dirty memory eviction lags behind.")),
-
-        sm::make_gauge(namestr +"_unspooled_dirty_bytes", [this] { return unspooled_dirty_memory(); },
-                       sm::description("Holds the size of used memory in bytes. Compare it to \"dirty_bytes\" to see how many memory is wasted (neither used nor available).")),
-    });
-}
-
 static const metrics::label class_label("class");
 
 void
@@ -1728,14 +1702,6 @@ future<mutation> database::do_apply_counter_update(column_family& cf, const froz
     });
 }
 
-future<> dirty_memory_manager::shutdown() {
-    _db_shutdown_requested = true;
-    _should_flush.signal();
-    return std::move(_waiting_flush).then([this] {
-        return _region_group.shutdown();
-    });
-}
-
 future<> memtable_list::flush() {
     if (!may_flush()) {
         return make_ready_future<>();
@@ -1766,77 +1732,6 @@ std::vector<replica::shared_memtable> memtable_list::clear_and_add() {
     std::vector<replica::shared_memtable> new_memtables;
     new_memtables.emplace_back(new_memtable());
     return std::exchange(_memtables, std::move(new_memtables));
-}
-
-future<flush_permit> flush_permit::reacquire_sstable_write_permit() && {
-    return _manager->get_flush_permit(std::move(_background_permit));
-}
-
-future<> dirty_memory_manager::flush_one(replica::memtable_list& mtlist, flush_permit&& permit) noexcept {
-    return mtlist.seal_active_memtable(std::move(permit)).handle_exception([this, schema = mtlist.back()->schema()] (std::exception_ptr ep) {
-        dblog.error("Failed to flush memtable, {}:{} - {}", schema->ks_name(), schema->cf_name(), ep);
-        return make_exception_future<>(ep);
-    });
-}
-
-future<> dirty_memory_manager::flush_when_needed() {
-    using namespace replica;
-    if (!_db) {
-        return make_ready_future<>();
-    }
-    // If there are explicit flushes requested, we must wait for them to finish before we stop.
-    return do_until([this] { return _db_shutdown_requested; }, [this] {
-        auto has_work = [this] { return has_pressure() || _db_shutdown_requested; };
-        return _should_flush.wait(std::move(has_work)).then([this] {
-            return get_flush_permit().then([this] (auto permit) {
-                // We give priority to explicit flushes. They are mainly user-initiated flushes,
-                // flushes coming from a DROP statement, or commitlog flushes.
-                if (_flush_serializer.waiters()) {
-                    return make_ready_future<>();
-                }
-                // condition abated while we waited for the semaphore
-                if (!this->has_pressure() || _db_shutdown_requested) {
-                    return make_ready_future<>();
-                }
-                // There are many criteria that can be used to select what is the best memtable to
-                // flush. Most of the time we want some coordination with the commitlog to allow us to
-                // release commitlog segments as early as we can.
-                //
-                // But during pressure condition, we'll just pick the CF that holds the largest
-                // memtable. The advantage of doing this is that this is objectively the one that will
-                // release the biggest amount of memory and is less likely to be generating tiny
-                // SSTables.
-                memtable& candidate_memtable = memtable::from_region(*(this->_region_group.get_largest_region()));
-                memtable_list& mtlist = *(candidate_memtable.get_memtable_list());
-
-                if (!candidate_memtable.region().evictable_occupancy()) {
-                    // Soft pressure, but nothing to flush. It could be due to fsync, memtable_to_cache lagging,
-                    // or candidate_memtable failed to flush.
-                    // Back off to avoid OOMing with flush continuations.
-                    return sleep(1ms);
-                }
-
-                // Do not wait. The semaphore will protect us against a concurrent flush. But we
-                // want to start a new one as soon as the permits are destroyed and the semaphore is
-                // made ready again, not when we are done with the current one.
-                (void)this->flush_one(mtlist, std::move(permit)).handle_exception([this] (std::exception_ptr ex) {
-                    dblog.error("Flushing memtable returned unexpected error: {}", ex);
-                });
-                return make_ready_future<>();
-            });
-        });
-    }).finally([this] {
-        // We'll try to acquire the permit here to make sure we only really stop when there are no
-        // in-flight flushes. Our stop condition checks for the presence of waiters, but it could be
-        // that we have no waiters, but a flush still in flight. We wait for all background work to
-        // stop. When that stops, we know that the foreground work in the _flush_serializer has
-        // stopped as well.
-        return get_units(_background_work_flush_serializer, _max_background_work).discard_result();
-    });
-}
-
-void dirty_memory_manager::start_reclaiming() noexcept {
-    _should_flush.signal();
 }
 
 future<> database::apply_in_memory(const frozen_mutation& m, schema_ptr m_schema, db::rp_handle&& h, db::timeout_clock::time_point timeout) {
